updated a ton of stuff; moved capstone project specific code to dataset.py in tools; DatabaseFactory (renamed from BuildDatabase) now fully leverages new io module; further abstracted configuration
