+++ tests/test_datastores.py
@@ -18,11 +18,10 @@ def test_SignalStore():
 
 def test_SignalGroup():
     rcd = get_data_record('30/3000063/3000063_0016', 'waveforms')
-    signals = {sig: SignalStore(rcd.p_signal[:, i], rcd.fmt[i]) for i, sig in enumerate(rcd.sig_name)}
+    signals = {sig: SignalStore(rcd.p_signal[:, i], rcd.fmt[i]) for i, sig in enumerate(rcd.sig_name) if sig in ['PLETH', 'ABP']}
     store = SignalGroup(signals)
     assert isinstance(store.ppg, SignalStore)
     assert isinstance(store.abp, SignalStore)
-    assert isinstance(store.ecg_ii, SignalStore)
 
 def test_WaveformRecord():
     tmp = get_data_record('30/3000063/3000063_0016', 'waveforms')
